# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Agricultural Nursery Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plants:
    get:
      operationId: getPlants
      tags: [plants]
      summary: Get all plants with optional filtering
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minHeight
          in: query
          schema:
            type: number
        - name: maxHeight
          in: query
          schema:
            type: number
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of plants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plant"

    post:
      operationId: createPlant
      tags: [plants]
      summary: Create a plant (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlantRequest"
      responses:
        "201":
          description: Plant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"
  /plants/{id}:
    get:
      operationId: getPlant
      tags: [plants]
      summary: Get a plant by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"
        "404":
          description: Plant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updatePlant
      tags: [plants]
      summary: Update a plant (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlantRequest"
      responses:
        "200":
          description: Plant updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"

    delete:
      operationId: deletePlant
      tags: [plants]
      summary: Delete a plant (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Plant deleted
  /plants/{id}/stock:
    patch:
      operationId: updatePlantStock
      tags: [plants]
      summary: Update plant stock (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStockRequest"
      responses:
        "200":
          description: Stock updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"
...
  /reports/summary:
    get:
      operationId: getReportSummary
      tags: [reports]
      summary: Get sales summary statistics (Admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Report summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportSummary"
  /reports/daily:
    get:
      operationId: getDailySales
      tags: [reports]
      summary: Get daily sales for the last 30 days (Admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Daily sales data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailySales"
  /reports/top-plants:
    get:
      operationId: getTopPlants
      tags: [reports]
      summary: Get most sold plants (Admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Top plants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopPlant"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    Category:
      type: object
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - nameAr
        - nameEn
        - createdAt
...
    Plant:
      type: object
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        categoryId:
          type: integer
          nullable: true
        category:
          $ref: "#/components/schemas/Category"
          nullable: true
        descriptionAr:
          type: string
          nullable: true
        descriptionEn:
          type: string
          nullable: true
        height:
          type: number
          nullable: true
        price:
          type: number
        stockQuantity:
          type: integer
        imageUrl:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
        featured:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - nameAr
        - nameEn
        - price
        - stockQuantity
        - featured
        - images
        - createdAt
        - updatedAt
    CreatePlantRequest:
      type: object
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        categoryId:
          type: integer
        descriptionAr:
          type: string
        descriptionEn:
          type: string
        height:
          type: number
        price:
          type: number
        stockQuantity:
          type: integer
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        featured:
          type: boolean
      required:
        - nameAr
        - nameEn
        - price
        - stockQuantity
    UpdateStockRequest:
      type: object
      properties:
        stockQuantity:
          type: integer
      required:
        - stockQuantity
...
    ReportSummary:
      type: object
      properties:
        totalRevenue:
          type: number
        todayRevenue:
          type: number
        monthRevenue:
          type: number
        totalOrders:
          type: number
        todayOrders:
          type: number
        monthOrders:
          type: number
        totalPlants:
          type: number
        lowStockPlants:
          type: number
        totalCustomers:
          type: number
      required:
        - totalRevenue
        - todayRevenue
        - monthRevenue
        - totalOrders
        - todayOrders
        - monthOrders
        - totalPlants
        - lowStockPlants
        - totalCustomers
    DailySales:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
        orders:
          type: integer
      required:
        - date
        - revenue
        - orders
    TopPlant:
      type: object
      properties:
        plantId:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        totalSold:
          type: integer
        totalRevenue:
          type: number
        imageUrl:
          type: string
          nullable: true
      required:
        - plantId
        - nameAr
        - nameEn
        - totalSold
        - totalRevenue